+++ daemon/src/GHCSpecter/Layouter/Packer.hs
@@ -1,9 +1,4 @@
-module GHCSpecter.Layouter.Box.Flow (
-  -- * shift
-  moveShapeBy,
-  moveBoundingBoxBy,
-  movePrimitiveBy,
-
+module GHCSpecter.Layouter.Packer (
   -- * flow
   toSizedLine,
   flowInline,
@@ -14,47 +9,15 @@ import Data.List qualified as L
 import Data.List.NonEmpty (NonEmpty)
 import Data.List.NonEmpty qualified as NE
 import GHCSpecter.Graphics.DSL (
-  DrawText (..),
-  Polyline (..),
   Primitive (..),
-  Rectangle (..),
-  Shape (..),
   ViewPort (..),
   getLeastUpperBoundingBox,
+  moveBoundingBoxBy,
+  movePrimitiveBy,
   viewPortHeight,
   viewPortSum,
   viewPortWidth,
  )
-
-moveShapeBy :: (Double, Double) -> Shape -> Shape
-moveShapeBy (dx, dy) (SDrawText (txt@DrawText {})) =
-  let (x, y) = dtextXY txt
-   in SDrawText (txt {dtextXY = (x + dx, y + dy)})
-moveShapeBy (dx, dy) (SPolyline (poly@Polyline {})) =
-  let (x0, y0) = plineStart poly
-      xs = plineBends poly
-      (x1, y1) = plineEnd poly
-      f (x, y) = (x + dx, y + dy)
-      poly' =
-        poly
-          { plineStart = f (x0, y0)
-          , plineBends = fmap f xs
-          , plineEnd = f (x1, y1)
-          }
-   in SPolyline poly'
-moveShapeBy (dx, dy) (SRectangle (rect@Rectangle {})) =
-  let (x, y) = rectXY rect
-   in SRectangle (rect {rectXY = (x + dx, y + dy)})
-
-moveBoundingBoxBy :: (Double, Double) -> ViewPort -> ViewPort
-moveBoundingBoxBy (dx, dy) (ViewPort (vx0, vy0) (vx1, vy1)) =
-  ViewPort (vx0 + dx, vy0 + dy) (vx1 + dx, vy1 + dy)
-
-movePrimitiveBy :: (Double, Double) -> Primitive e -> Primitive e
-movePrimitiveBy (dx, dy) (Primitive shape vp hitEvent) =
-  let shape' = moveShapeBy (dx, dy) shape
-      vp' = moveBoundingBoxBy (dx, dy) vp
-   in Primitive shape' vp' hitEvent
 
 toSizedLine :: NonEmpty (Primitive a) -> (ViewPort, NonEmpty (Primitive a))
 toSizedLine xs = (getLeastUpperBoundingBox xs, xs)
