+++ daemon/src/GHCSpecter/UI/Console.hs
@@ -34,7 +34,7 @@ import GHCSpecter.Graphics.DSL (
   rectangle,
   viewPortHeight,
  )
-import GHCSpecter.Layouter.Box.Flow (
+import GHCSpecter.Layouter.Packer (
   flowInline,
   flowLineByLine,
   toSizedLine,
